!!! TASK: Remove CheckpointStorageInterface
Removes the `CheckpointStorageInterface` and the `CatchUp` class (that relied upon it) from this package in order to keep this package focused on its main feature set.

For 3rd parties that relied on this code, the [CatchUp](https://github.com/neos/eventstore/blob/bede780bd4e5f4d743ec6a9d4d9546df3c89206a/src/CatchUp/CatchUp.php) helper should be re-implemented for now.
In the midterm, we plan to create a higher level package for projection and replay logic

Resolves: #18