!!! FEATURE: Make causation and correlation id an actual concept
Previously, `causationId` and `correlationId` were just implicitly
supported via `EventMetaData`.
With this change, the `Event` model contains (optional) fields for
both.

This is a potentially breaking change, because it changes the signature of `Event` constructor (making `metadata` nullable, adding `causationId` and `correlationId` arguments).

Furthermore this adjusts the signature of `EventStoreInterface::commit()` to support committing of single events.